Node can decide whether it can be deleted

+++ egui_node_graph/src/editor_ui.rs
@@ -771,7 +771,13 @@ where
         // --- Interaction ---
 
         // Titlebar buttons
-        if Self::close_button(ui, outer_rect).clicked() {
+        let can_delete = self.graph.nodes[self.node_id].user_data.can_delete(
+            self.node_id,
+            self.graph,
+            user_state,
+        );
+
+        if can_delete && Self::close_button(ui, outer_rect).clicked() {
             responses.push(NodeResponse::DeleteNodeUi(self.node_id));
         };
